requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiokafka==0.8.1
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.2
anyio==4.0.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.4
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.3
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.3
click-params==0.3.0
cloudpickle==2.2.1
cmdstanpy==1.2.0
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
Cython==3.0.4
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.89.1
fastapi-utils==0.2.1
Flask==2.3.3
fonttools==4.43.1
frozenlist==1.4.0
gevent==23.9.1
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.40
greenlet==3.0.0
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
holidays==0.35
httplib2==0.19.1
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipinfo==4.4.3
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
kafka-python==2.0.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mlflow==2.6.0
mlserver==1.3.5
mlserver-mlflow==1.3.5
multidict==6.0.4
mysql-connector-python==8.0.33
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
orjson==3.8.14
packaging==23.2
pandas==2.1.1
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prophet==1.1.5
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-grpc-prometheus==0.7.0
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.7.0
Pympler==1.0.1
PyMySQL==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-rapidjson==1.12
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
semver==3.0.2
six==1.16.0
sklearn==0.0.post10
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
sqlparse==0.4.4
stack-data==0.6.3
stanio==0.3.0
starlette==0.22.0
starlette-exporter==0.16.0
statsmodels==0.14.0
streamlit==1.12.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
tritonclient==2.36.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.18.0
validators==0.18.2
watchfiles==0.21.0
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
yarl==1.9.2
zenml==0.45.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
